requirements.txt
# Get same packages here as environment.yml and the right python package with docker FROM
azure-storage-blob==12.11.0
azure-storage-file-share==12.7.0
azure-storage-file==2.1.0
azure-data-tables==12.0.0  
azure.identity==1.7.0
requests==2.27.1
joblib==1.1.0  